# fix: enhance toolchain setup process and feedback mechanism