To resolve #2074 by reusing the `SetupProto` and `SetupToolchain` actions.

## Screenshots

New: 
```sh
moon on  opt/setup-with-new-action [!] via  v24.5.0 via 🦀 v1.88.0 
❯ cargo run --bin moon setup
    Finished `dev` profile [unoptimized] target(s) in 0.60s
     Running `target/debug/moon setup`
▪▪▪▪ installing proto 0.51.4
▪▪▪▪ installing bun 1.2.2
▪▪▪▪ installing deno 2.1.9
▪▪▪▪ installing python 3.11.10
▪▪▪▪ installing node 22.14.0
▪▪▪▪ installing yarn 4.8.0

Tasks: 0 tasks ran
 Time: 28s 961ms ❯❯❯❯ to the moon


│ SUCCESS
│ Successfully setup 5 toolchain(s)!
```

Original with issue:
```sh
moon on  opt/setup-with-new-action [!] via  v24.5.0 via 🦀 v1.88.0 
❯ moon setup
━━━━━━━━╾            Downloading and installing toolchains...
━━━━━━━━━━━━━━━━━━━━ Downloading and installing toolchains...
Error: proto::store::shim::missing_binary

  × Unable to create shims as the proto-shim binary cannot be found.
  │ Looked in the PROTO_HOME environment variable and ~/.proto/bin directory.
```

Original with success(workaround):
```
moon on  opt/setup-with-new-action [!] via  v24.5.0 via 🦀 v1.88.0 took 30s 
❯ moon setup

│ SUCCESS
│ Toolchains have been setup!
```